openNewGameScreen now creates a MenuScreen, DifficultyScreen is now created with a new function: openNewDifficultyScreen
Created textures for the Main and Help menus

Created a new HelpScreen

Created 2 enums for the MenuScreen and HelpScreen: Menu and Help

Started working on a main menu, it's currently able to start the game
and quit the game.

Destroying the friendly college (via ramming) now results in a game
over.

LossScreen now has a unique constructor with a String, to allow for
custom loss messages.