8/4 Shop Branch Commit 1:
Began work on ShopScreen, added 6 purchases so far.

Some are one-off purchases and will be disabled after purchasing.

Skipping the Storm is situational; it cannot be purchased if there is no
storm.

PirateGame now retains the current GameScreen and ShopScreen to allow
switching back and forth between the two.

GameScreen now has an addPurchase and the EntityShip now has an
applyPurchase function to integrate shop purchases.

EntityShip now has some extra stat multipliers to match the
corresponding shop purchase.

Tweaked the GameScreen show() function to not call certain events
intended for the start of the game multiple times.